First round update to 1.6.4

+++ RpgEssentials/src/com/topcat/npclib/entity/NPC.java
@@ -3,8 +3,8 @@ package com.topcat.npclib.entity;
 import java.util.ArrayList;
 import java.util.Iterator;
 
-import net.minecraft.server.v1_4_R1.Entity;
-import net.minecraft.server.v1_4_R1.EntityPlayer;
+import net.minecraft.server.v1_6_R3.Entity;
+import net.minecraft.server.v1_6_R3.EntityPlayer;
 
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
@@ -122,7 +122,7 @@ public class NPC {
 						look = (float) (Math.toDegrees(Math.asin(last.b.getY() - n.b.getY())) / 2);
 					}
 					getEntity().setPositionRotation(n.b.getX() + 0.5, n.b.getY(), n.b.getZ() + 0.5, angle, look);
-					((EntityPlayer)getEntity()).az = angle;//ay
+					((EntityPlayer)getEntity()).az = (int) angle;//ay
 				} else {
 					onFail.run();
 				}
@@ -130,7 +130,7 @@ public class NPC {
 			last = n;
 		} else {
 			getEntity().setPositionRotation(runningPath.getEnd().getX(), runningPath.getEnd().getY(), runningPath.getEnd().getZ(), runningPath.getEnd().getYaw(), runningPath.getEnd().getPitch());
-			((EntityPlayer)getEntity()).az = runningPath.getEnd().getYaw();//ay
+			((EntityPlayer)getEntity()).az = (int) runningPath.getEnd().getYaw();//ay
 			Bukkit.getServer().getScheduler().cancelTask(taskid);
 			taskid = 0;
 		}
